refactor: Split out repetition parsers to a module
